SpeechRecognizerCore/pyproject.toml
[tool.poetry]
name = "speech-recognizer-core"
version = "0.1.0"
description = "Core NSTT"
readme = "README.md"
authors = [
    "Vitalii Dolhushyn <vitalii.dolhushyn@viti.edu.ua>",
    "Borys Ohorodnii <borys.ohorodnii@viti.edu.ua>",
    "Dmytro Bykov <dmytro.bykov@viti.edu.ua>",
    "Oleksandr Zahorenko <oleksandr.zahorenko@viti.edu.ua>",
    "Artem Voznyi <artem.voznyi@viti.edu.ua>",
]
packages = [
    { include = "source_recognizer" },
    { include = "converter_to_wav" },
    { include = "denoise" },
    { include = "language_identification" },
    { include = "speech_recognizer" },
]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
urllib3 = "2.1.0"
certifi = "2023.11.17"
idna = "3.6"
charset-normalizer = "3.3.2"
requests = "2.31.0"
typing-extensions = "4.9.0"
pysndfx = "^0.3.6"
keyboard = "0.13.5"
pyaudio = "0.2.12"
librosa = "0.10.1"
pydub = "^0.25.1"
soundfile = "^0.12.1"
speechrecognition = "^3.10.1"
tensorflow = "2.15.0"
keras = "2.15.0"
numpy = "1.26.3"
scipy = "^1.12.0"
loguru = "0.7.2"
noisereduce = "^3.0.0"
torch = "^2.2.0"
torchvision = "0.17.0"
tensorflow-io-gcs-filesystem = "0.31.0"
eel = "^0.16.0"
tk = "^0.1.0"
azure-storage-blob = "^12.19.0"
pyannote-audio = "^3.1.1"
ttkthemes = "^3.2.2"
matplotlib = "3.8.2"
pillow = "10.2.0"
pocketsphinx = "^5.0.3"
openai-whisper = "^20231117"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "main:main"
